Move truecommand service to service parts
